instaweb: Make all log files available in a single zip file
At times it's desirable to make log files of a test execution available
outside a `stbt batch` report because
* they should be sent to someone who doesn't have access to the server
  the test results are stored on;
* they should be attached to a bug report;
* one would like to use their preferred text editor to view log files.

In these cases it's quite inconvenient to download logs one-by-one from
the web interface; while granting ssh or ftp access to the web server
for all users of the test system is insecure.

This commit adds the link "[Download all]" to the bottom of the "Logs"
section of a test result's page. It requests a single zip file that
contains all logs that belong to the test result. To save hard drive
space, the package is generated on request. "zip" file format is
preferred because it's supported by default in all operating systems.