stbt-batch.d/run: Add a config option to disable saving video
Rebooting the set-top-box with `--save-video` still frequently fails
with GStreamer error or NoVideo. Until this issue is resolved, we'd like
to work around the failures in tests in which we don't want to watch the
recorded video, simply by disabling the `--save-video` option.

To disable `--save-video`, add

    [batch]
    save_video = False

to the stbt config file. Defaults to True.