stbt.py: Create a new video file each time video recording is restarted
The previous commit modified stbt.py to restart the source pipeline
every time the source pipeline is restarted. It required us to use
filesink in "append" mode to avoid overwriting the video file after each
restart.

Unfortunately the video saved this way is not seekable. A workaround
for this issue is to save video to a new file after every video encoder
restart. The resulting files are all error-free and seekable.

Running
    stbt run --save-video video.webm test.py
with a test that involves 2 source pipeline restarts will create
3 files: video.webm, video1.webm and video2.webm.