stbt.py: Make decodebin2 handle asynchronous state changes
Even with the fix "Stop sink pipeline while the source pipeline is
restarting", we get occasional GStreamer errors during source pipeline
restarts. decodebin2 raises "The stream is of a different type than
handled by this element.: No valid frames found before end of stream".

The solution is to set the "async-handling" parameter of decodebin2 to
true. As decodebin2 is in the user-defined source pipeline, we have to
check whether this element exists and set the parameter dynamically.