stbt.py: Avoid race condition in restarting source pipeline
stb-tester doesn't restart the source pipeline immediately after
receiving an EOS message; it starts a 5-second timer. However, if it
receives another EOS message from the source bus within 5 seconds, it
has two competing restart procedures.

As the first procedure deletes the source pipeline (sets it to None),
the second one raises "AttributeError: 'NoneType' object has no
attribute 'set_state'". This error is not raised in the main thread and
therefore it has no effect, but it's confusing to see in the log file.