stbt batch run: Work around VirtualBox issue 10085
VirtualBox has a long standing bug* that creating a symlink within a
folder shared with a Windows host fails with "Protocol error". As
`stbt batch run` creates the `current` symlink as part of the set-up
process, the test run fails with the error "Failed to set up test-run
directory".

To work around this issue, do not raise error if creating the `current`
symlink fails. It is created for convenience and `stbt batch` is fully
usable without it.

* https://www.virtualbox.org/ticket/10085