stbt.py: Save a streamable video and regenerate EBML header at the end
The previous commit modified stbt.py to stop and restart the sink
pipeline every time the source pipeline is rebuilt.

Unfortunately it has an issue that the video file is corrupt because the
video file index is recreated for each file segment. A solution is to
create a `seekable` file that doesn't contain a video file index, and
rebuild the index at the end.

This commit introduces a new dependency, `ffmpeg` that we use to
recreate video file index.