stbt.py: [bug] Save temporary video in the original video's directory
Commit 4d2a9199 modified `stbt.teardown` to regenerate the header of the
saved video file as part of the teardown process. It generates the video
to $PWD/.seekable.webm and then replaces the original video file with
it.

Firstly, if the current working directory is on a different partition
from the original video file, `os.rename` raises "Invalid cross-device
link" error. Secondly, we should make sure that the temporary file has
a unique name to avoid the unlikely event that if two video files are
being generated at the same time in the same directory then they
overwrite each other.